Support 64-bit integers for POS and INFO/FORMAT fields
int64 fields now flow through reading, filtering, querying and region
selection at native width (these paths were already width-agnostic). The
C VCF text encoder only supports 1-, 2- and 4-byte integers, so a new
utils.to_vcf_int32 casts int64 down to int32 at the encoder boundary and
raises a clear error when a value falls outside the int32 range; it is
applied to POS and to int64 INFO/FORMAT fields in VcfWriter.write_chunk
(replacing the silent POS truncation). The fixed-size bgen encoder gets
the same overflow guard; plink writes POS as text and is already
int64-safe.

Also fix variant_N_ALT to compute int64 to match its declared dtype.

Testing: the dtype matrix gains an int64 token (ID fields), extending
read-back/header/view/query/filter coverage to int64 for in-range
values. New focused tests cover the overflow-raises behaviour, large
int64 INFO values through filter/query, and int64 POS end-to-end
(reader, region filter, query) including the bgen guard and plink
text round-trip. vcz_builder.make_vcz gains position_dtype to build
64-bit-position fixtures.

+++ vcztools/virtual_fields.py
@@ -91,7 +91,7 @@ def _compute_ns(deps, cache):
 
 def _compute_n_alt(deps, cache):
     alt = np.asarray(deps["variant_allele"])[:, 1:]
-    return (alt != "").sum(axis=1).astype(np.int32)
+    return (alt != "").sum(axis=1).astype(np.int64)
 
 
 def _compute_n_missing(deps, cache):
